requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
psycopg[binary]>=3.1.0
djangorestframework-simplejwt>=5.3.1
django-filter>=23.5
django-cors-headers>=4.3.1
python-decouple>=3.8
drf-spectacular>=0.27.0
